Make lazy: move
