+++ src/CMakeLists.txt
@@ -20,6 +20,7 @@ set(YACLIB_SOURCES
 
 add_subdirectory(algo)
 add_subdirectory(async)
+add_subdirectory(lazy)
 add_subdirectory(exe)
 add_subdirectory(util)
 add_subdirectory(fault)
