+++ src/async/CMakeLists.txt
@@ -3,7 +3,6 @@ list(APPEND YACLIB_INCLUDES
   ${YACLIB_INCLUDE_DIR}/async/future.hpp
   ${YACLIB_INCLUDE_DIR}/async/promise.hpp
   ${YACLIB_INCLUDE_DIR}/async/run.hpp
-  ${YACLIB_INCLUDE_DIR}/async/task.hpp
   )
 list(APPEND YACLIB_HEADERS
   ${YACLIB_INCLUDE_DIR}/async/detail/base_core.hpp
@@ -18,7 +17,6 @@ list(APPEND YACLIB_SOURCES
   ${CMAKE_CURRENT_SOURCE_DIR}/future_impl.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/promise_impl.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/result_core.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/task_impl.cpp
   )
 
 add_files()
